Create pipelines list page
