+++ frontend/public/components/default-resource.jsx
@@ -127,7 +127,7 @@ export const DefaultPage = (props) => (
   <ListPage
     {...props}
     ListComponent={DefaultList}
-    canCreate={props.canCreate === undefined ? _.get(kindObj(props.kind), 'crd') : props.canCreate}
+    canCreate={props.canCreate ?? _.get(kindObj(props.kind), 'crd')}
   />
 );
 DefaultPage.displayName = 'DefaultPage';
